requirements.txt
apache-airflow-providers-duckdb